update makefile to work with anaconda envs
Locations of Python dependencies were hardcoded in for system python,
specifically, system python on Ubuntu.  Now it looks for the location of
the Python binary on $PATH and then uses that location as a basis for
determining the relative locations of the necessary libraries.

Tested on Arch and Ubuntu using both Anaconda and system Python.