Update README - last dependencies tested
I replace the last version of each dependency that I tested on Blackbird. I was wondering if in the installation instructions for python, numpy and swig we should change them for a `conda install` version, because the default versions that come with Ubuntu for example, are usually older than the ones from Anaconda. 

@gilforsyth and @cdcooper what do you think?