change handling of system arguments
instead of a script, the main body is now wrapped in a `main()`.  But we
would like to be able to pass command-line arguments to the script, but
also to debug it or use it interactively.

So!

`main()` takes the parameter `argv` which can be passed in by the
user (within, say IPython), or, if the script is run directly (assumed
standard usage), it passes in `sys.argv`


Former-commit-id: 146e6162e0928c0b42d4c199aa8b6004cb1d2bc6