Renaming matrix_test to pygbe_matrix and moving it out from tests
This folder contains a pygbe matrix version that is completely
independent from pygbe, so it shouldn't be inside the pygbe folder.
It's a small version of pygbe that we use to test some cases.
Actually inside the folder test that lives in pygbe we have a
script called run_matrix_tests that compares the outputs
from pygbe and pygbe-matrix.