remove which.py and add check for nvcc version
I removed the pure Python implementation of `which` that I added in
since it seems like unnecessary bloat. The only usage of it is to check
that `nvcc` is on path and that can be done with a `subprocess` call
instead.

I also added a check to ensure that the version of `nvcc` being used is
<= 7.0


Former-commit-id: fd83d73cc00b00ea69952b42d961d8bfeea6c549