unify import behavior and prepare for yapf format run
I did the following:

1. Stripped out a few remaining `from numpy import *`
2. Removed all remaining `sys.path` hacking and instead switched these
to use proper module importing
3. All `from math import pi` now use `from numpy import pi`

4. Added `yapf enable` and `yapf disable` flags where I thought they
were needed.

Explanation:
YAPF is very good at formatting code to follow PEP8 and it will save us
a LOT of work getting PyGBe to match the existing standard, but there
are certain parts of PyGBe which likely work better as they are, in
terms of readability, etc...

Specifically I am thinking of `pygbe/classes.py` where all of the
quadratureRules are defined.  While the current format may not be PEP8
compliant, it makes sense to me to have the various quadrature points
formatted the way that they are, so I have added in flags to prevent
YAPF from applying any formatting changes to these (and in a few other
places).  In addition, I manually fixed up these exempted sections with
regard to trailing whitespace, consistency of spacing, removal of tabs,
etc... so everything should be good to go