refactor 1 set of reg. tests into problem folder format
As a P.O.C., the molecule_dirichlet reg. test now works as follows:

3 folders have been created in regression_tests:
molecule_dirichlet/
molecule_single_center/
dirichlet_surface/

each folder has it's own .config files (which are tiny) and symlinks to
the *.phi0 files (which may be followed or copied by git, not sure
which).

geometry is handled by means of a new flag in pygbe.main.main that
specifies a custom geometry prefix (so we don't have to have any copies
of the mesh lying around)