reformat all regression tests to use new framework
results by default are cached in a pickled dictionary so that the same
case isn't run repeatedly and instead uses the already calculated result
to speed up the time the regression tests take.  Also eliminates most of
the repeated code that was present.

for now all plotting stuff is commented out while we decide if we still
want it


Former-commit-id: 82ea4d8e45414a785c8bd191e587c72005a23f0a