fix bug with output file not being closed at end of main
I swear I already fixed this but apparently I didn't... Maybe didn't
push?  Didn't commit?  No idea.  I'm confused.

In any case:
The regression tests read PyGBe's output from the logfiles and search
through them for the appropriate values to check (this itself is less
than ideal, but functional).  However, since we redirect stdout in order
to write to both the file and the screen, if we don't STOP doing that at
the end of a PyGBe run, stdout is still a mix of screen/logfile.  And
since it's still active, the file that we're writing to is still open.

If the file is still opened by a process and you try to open it
elsewhere, it will appear empty, since things aren't dumped to file
until the objects that have a lock on them are closed.  So we search for
numbers in an empty file.

Oops.  This resets stdout to it's regular state at the end of the main
function in PyGBe, thus closing the output file, ensuring it's written
and available before the regression tests look for values in it.


Former-commit-id: 1ed91b24336e300b4abd4f4b29394de882dd5036