Improve regression test logic
There was a particularly silly bit of code in here (that I wrote) that
used an if statement to check whether or not to take the abs() of some
values. That has been removed since it was useless.

The only caveat is that the two_spheres problem sometimes has zero
values, so I've also added a 1e-16 to the denominator of that test so we
don't get a NaN.