add versioneer
This adds versioneer to PyGBe.  versioneer is a small python package (no
extra install required) that automatically updates version numbers.
After this is merged into master, you will be able to query
`pygbe.__version__` and it will respond with a tag number and also the
git hash of the repo when the install was performed.

It's a way to automatically track the exact version of PyGBe that was
used for some purpose.

Once this is in, we can start including that version info in results
dictionaries, so we'll have the date and time the run was performed,
various other essential information and also the exact hash of PyGBe
that ran it.