first stage of new logging framework
Using the builtin ``logging`` module. Any logger that's created in any module
will belong to the "root" logger, which is created in `pygbe/main.py`.

All of the messages will be in that logger and written to an output file, with
accompanying info about where that info/error, etc was created.

Ideally, we can add custom message types to the log and then determine what gets
written to screen based on levels determined by keyword arguments (so when using
PyGBe programmatically we don't need to see the iteration counts, for example)