Update Docker -  pytest not compatible with pycuda
When installing pycuda it looks for a version of pytest >= 2, as right now it pulls pytest 4.0 but it doesn't find it because of a problem in the installation of pytest. Error message:
Installed /opt/miniconda/lib/python3.6/site-packages/UNKNOWN-4.0.0-py3.6.egg
error: The 'pytest>=2' distribution was not found and is required by pycuda

Problem solved by installing previous pycuda installation, an older version of pytest.